OCLOMRS-1126: Code refactor
