+++ api/src/main/java/org/openmrs/module/openconceptlab/client/OclConcept.java
@@ -203,14 +203,7 @@ public class OclConcept {
 	}
 
 	@JsonIgnoreProperties(ignoreUnknown = true)
-	public static class Name {
-
-		private String uuid;
-
-		@JsonProperty("external_id")
-		private String externalId;
-
-		private Locale locale;
+	public static class Name extends OclLocalizedObject {
 
 		@JsonProperty("locale_preferred")
 		private boolean localePreferred;
@@ -221,35 +214,6 @@ public class OclConcept {
 		private String nameType;
 
 		private String type;
-
-		private Boolean retired;
-
-		@JsonProperty("retire_reason")
-		private String retireReason;
-
-		public String getUuid() {
-			return uuid;
-		}
-
-		public void setUuid(String uuid) {
-			this.uuid = uuid;
-		}
-
-		public String getExternalId() {
-			return externalId;
-		}
-
-		public void setExternalId(String externalId) {
-			this.externalId = externalId;
-		}
-
-		public Locale getLocale() {
-			return locale;
-		}
-
-		public void setLocale(Locale locale) {
-			this.locale = locale;
-		}
 
 		public boolean isLocalePreferred() {
 			return localePreferred;
@@ -283,26 +247,6 @@ public class OclConcept {
 			this.type = type;
 		}
 
-		public Boolean getRetired() {
-			return retired;
-		}
-
-		public boolean isRetired() {
-			return Boolean.TRUE.equals(retired);
-		}
-
-		public void setRetired(Boolean retired) {
-			this.retired = retired;
-		}
-
-		public String getRetireReason() {
-			return retireReason;
-		}
-
-		public void setRetireReason(String retireReason) {
-			this.retireReason = retireReason;
-		}
-
 		@Override
 		public String toString() {
 			return new ToStringBuilder(this).append("name", name).build();
@@ -310,7 +254,7 @@ public class OclConcept {
 
 		@Override
 		public int hashCode() {
-			return new HashCodeBuilder().append(locale).append(localePreferred).append(name).append(nameType).build();
+			return new HashCodeBuilder().append(getLocale()).append(localePreferred).append(name).append(nameType).build();
 		}
 
 		@Override
@@ -325,58 +269,22 @@ public class OclConcept {
 				return false;
 			}
 			Name rhs = (Name) obj;
-			return new EqualsBuilder().append(locale, rhs.locale).append(localePreferred, rhs.localePreferred)
+			return new EqualsBuilder().append(getLocale(), rhs.getLocale()).append(localePreferred, rhs.localePreferred)
 			        .append(name, rhs.name).append(nameType, rhs.nameType).build();
 		}
 
 		public void copyFrom(ConceptName name) {
 			this.name = name.getName();
-			locale = name.getLocale();
+			setLocale(name.getLocale());
 			localePreferred = name.getLocalePreferred() != null ? name.getLocalePreferred() : false;
 			nameType = name.getConceptNameType() != null ? name.getConceptNameType().toString() : null;
 		}
 	}
 
 	@JsonIgnoreProperties(ignoreUnknown = true)
-	public static class Description {
-
-		private String uuid;
-
-		@JsonProperty("external_id")
-		private String externalId;
-
-		private Locale locale;
+	public static class Description extends OclLocalizedObject {
 
 		private String description;
-
-		private Boolean retired;
-
-		@JsonProperty("retire_reason")
-		private String retireReason;
-
-		public String getUuid() {
-			return uuid;
-		}
-
-		public void setUuid(String uuid) {
-			this.uuid = uuid;
-		}
-
-		public String getExternalId() {
-			return externalId;
-		}
-
-		public void setExternalId(String externalId) {
-			this.externalId = externalId;
-		}
-
-		public Locale getLocale() {
-			return locale;
-		}
-
-		public void setLocale(Locale locale) {
-			this.locale = locale;
-		}
 
 		public String getDescription() {
 			return description;
@@ -386,34 +294,14 @@ public class OclConcept {
 			this.description = description;
 		}
 
-		public Boolean getRetired() {
-			return retired;
-		}
-
-		public boolean isRetired() {
-			return Boolean.TRUE.equals(retired);
-		}
-
-		public void setRetired(Boolean retired) {
-			this.retired = retired;
-		}
-
-		public String getRetireReason() {
-			return retireReason;
-		}
-
-		public void setRetireReason(String retireReason) {
-			this.retireReason = retireReason;
-		}
-
 		public void copyFrom(ConceptDescription description) {
 			this.description = description.getDescription();
-			locale = description.getLocale();
+			setLocale(description.getLocale());
 		}
 
 		@Override
 		public int hashCode() {
-			return new HashCodeBuilder().append(locale).append(description).build();
+			return new HashCodeBuilder().append(getLocale()).append(description).build();
 		}
 
 		@Override
@@ -428,12 +316,12 @@ public class OclConcept {
 				return false;
 			}
 			Description rhs = (Description) obj;
-			return new EqualsBuilder().append(locale, rhs.locale).append(description, rhs.description).build();
+			return new EqualsBuilder().append(getLocale(), rhs.getLocale()).append(description, rhs.description).build();
 		}
 
 		@Override
 		public String toString() {
-			return new ToStringBuilder(this).append("description", description).append("locale", locale).build();
+			return new ToStringBuilder(this).append("description", description).append("locale", getLocale()).build();
 		}
 	}
