+++ api/src/main/java/org/openmrs/module/openconceptlab/client/OclLocalizedObject.java
@@ -1,0 +1,73 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ *
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.openconceptlab.client;
+
+import java.util.Locale;
+
+import org.codehaus.jackson.annotate.JsonProperty;
+
+public abstract class OclLocalizedObject {
+
+	private String uuid;
+
+	@JsonProperty("external_id")
+	private String externalId;
+
+	private Locale locale;
+
+	private Boolean retired;
+
+	@JsonProperty("retire_reason")
+	private String retireReason;
+
+	public String getUuid() {
+		return uuid;
+	}
+
+	public void setUuid(String uuid) {
+		this.uuid = uuid;
+	}
+
+	public String getExternalId() {
+		return externalId;
+	}
+
+	public void setExternalId(String externalId) {
+		this.externalId = externalId;
+	}
+
+	public Locale getLocale() {
+		return locale;
+	}
+
+	public void setLocale(Locale locale) {
+		this.locale = locale;
+	}
+
+	public Boolean getRetired() {
+		return retired;
+	}
+
+	public boolean isRetired() {
+		return Boolean.TRUE.equals(retired);
+	}
+
+	public void setRetired(Boolean retired) {
+		this.retired = retired;
+	}
+
+	public String getRetireReason() {
+		return retireReason;
+	}
+
+	public void setRetireReason(String retireReason) {
+		this.retireReason = retireReason;
+	}
+}
